+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## Next Version
+### Features
+* Show warning if there are textures that are crunch compressed.
+* Add list of all crunch compressed textures to the debug info section.
+
 ### Bug fixes
 * Fixed the delete unused gameobjects function not respecting the exclusions properly.
-  
+* Don't merge textures that are crunch compressed. Fix for: [Material merging results in corrupted material/textures](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/5)
+
 ### Debug info
 * Add original shader name to the generated shader as a comment.